feat(Redis优化): 完成点赞/收藏高并发优化 + 压测 + 文档
- 添加收藏功能 nil 检查，避免 panic
- 压测脚本：复用 HTTP 连接池，支持 4 个场景测试
- 性能指标：点赞 1395 QPS，取消点赞 1974 QPS，收藏 1208 QPS
- 完整文档：架构设计、代码实现、性能测试、简历素材

+++ flashvid-platform-gin/internal/task/sync_stats.go
@@ -73,7 +73,7 @@ func syncOnce(ctx context.Context) error {
 		// 更新 MySQL（直接覆盖，因为 Redis 已是最新值）
 		_, err = query.Video.WithContext(ctx).
 			Where(query.Video.ID.Eq(videoID)).
-			UpdateSimple(query.Video.LikeCount.Value(likeCount))
+			UpdateSimple(query.Video.LikeCount.Value(int32(likeCount)))
 		if err != nil {
 			fmt.Printf("[SyncVideoStats] update video %d failed: %v\n", videoID, err)
 			continue
